backend/requirements.txt
Django>=4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
python-dotenv==1.0.0 # For managing environment variables
psycopg2==2.9.7 # PostgreSQL adapter for Python
plaid-python==28.0.0 # Plaid API client for Python
requests==2.31.0 # HTTP library for Python